+++ activemq-client/src/main/java/org/apache/activemq/openwire/OpenWireFormat.java
@@ -532,13 +532,13 @@ public final class OpenWireFormat implements WireFormat {
         // We can only cache that item if there is space left.
         if (marshallCacheMap.size() < marshallCache.length) {
             marshallCache[i] = o;
-            Short index = Short.valueOf(i);
+            Short index = i;
             marshallCacheMap.put(o, index);
             return index;
         } else {
             // Use -1 to indicate that the value was not cached due to cache
             // being full.
-            return Short.valueOf((short)-1);
+            return (short) -1;
         }
     }
