+++ activemq-kahadb-store/src/main/java/org/apache/activemq/store/kahadb/disk/journal/Journal.java
@@ -718,7 +718,7 @@ public class Journal {
     }
 
     DataFile getDataFile(Location item) throws IOException {
-        Integer key = Integer.valueOf(item.getDataFileId());
+        Integer key = item.getDataFileId();
         DataFile dataFile = null;
         synchronized (currentDataFile) {
             dataFile = fileMap.get(key);
@@ -906,7 +906,7 @@ public class Journal {
                 if (dataFile == null) {
                     return null;
                 } else {
-                    cur.setDataFileId(dataFile.getDataFileId().intValue());
+                    cur.setDataFileId(dataFile.getDataFileId());
                     cur.setOffset(0);
                     if (limit != null && cur.compareTo(limit) >= 0) {
                         LOG.trace("reached limit: {} at: {}", limit, cur);
@@ -1045,7 +1045,7 @@ public class Journal {
 
     public DataFile getDataFileById(int dataFileId) {
         synchronized (currentDataFile) {
-            return fileMap.get(Integer.valueOf(dataFileId));
+            return fileMap.get(dataFileId);
         }
     }
