+++ activemq-tooling/activemq-memtest-maven-plugin/src/main/java/org/apache/activemq/tool/MemoryMonitoringTool.java
@@ -67,7 +67,7 @@ public class MemoryMonitoringTool implements Runnable {
     public Thread startMonitor() {
 
         String intervalStr = this.getTestSettings().getProperty("checkpoint_interval");
-        checkpointInterval = Integer.valueOf(intervalStr).intValue();
+        checkpointInterval = Integer.valueOf(intervalStr);
         this.getTestSettings().remove("checkpoint_interval");
 
         memoryBean = ManagementFactory.getMemoryMXBean();
